+++ crates/features/src/input/dtos/move_dto.rs
@@ -1,8 +1,13 @@
 use nest_rs::core::input;
 
+use crate::screen::RegionDto;
+
 #[input]
 #[derive(Debug)]
 pub struct MoveDto {
     pub x: i64,
     pub y: i64,
+
+    #[serde(default)]
+    pub region: Option<RegionDto>,
 }